Validation regex added
